requirements.txt
django-allauth
django-allauth-ui
django-widget-tweaks